Finalized version of user-end reporting flow.
The user can initiate a reporting flow by entering the  keyword. Once the bot is active,
the user can supply the link to a message that they would like to report, which the bot will
then fetch and prompt the user to indicate what type of abuse it is. From there, the user selects
an abuse type and is then reprompted with subcategories under the original type of abuse, and is given
the opportunity to provide additional context with the report and block the offending poster.